Add AmbiguousTimeHandling
By default we still raise TZInfo::AmbiguousTime, however, we now also expose
the ability to conveniently resolve ambiguous local time issues by yielding
the possible periods and the ambiguous local time. For convenience we provide
a handler that defaults to the DST period and a handler that defaults to the
standard time period.